# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rugby Manager API",
    "version": "1.0.0",
    "description": "API para gestión de equipos de rugby - Jugadores, Partidos, Entrenamientos y más"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de desarrollo"
    }
  ],
  "components": {
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id_usuario": {
            "type": "string",
            "format": "uuid"
          },
          "correo": {
            "type": "string",
            "format": "email",
            "example": "juan@ejemplo.com"
          },
          "apellido": {
            "type": "string",
            "example": "García"
          },
          "nombre": {
            "type": "string",
            "example": "Juan"
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UsuarioUpdate": {
        "type": "object",
        "properties": {
          "correo": {
            "type": "string",
            "format": "email"
          },
          "apellido": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/calendar/equipo/{idEquipo}": {
      "get": {
        "summary": "Get calendar events (matches + trainings) for a team",
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idEquipo",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (YYYY-MM-DD), defaults to first of current month"
          },
          {
            "in": "query",
            "name": "hasta",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (YYYY-MM-DD), defaults to last of current month"
          }
        ],
        "responses": {
          "200": {
            "description": "Calendar events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "desde": {
                      "type": "string",
                      "format": "date"
                    },
                    "hasta": {
                      "type": "string",
                      "format": "date"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "PARTIDO",
                              "ENTRENAMIENTO"
                            ]
                          },
                          "title": {
                            "type": "string"
                          },
                          "start": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "end": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "color": {
                            "type": "string"
                          },
                          "data": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invitaciones": {
      "post": {
        "summary": "Create invitation code",
        "tags": [
          "Invitations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id_equipo",
                  "rol",
                  "id_creador"
                ],
                "properties": {
                  "id_equipo": {
                    "type": "string"
                  },
                  "rol": {
                    "type": "string",
                    "enum": [
                      "ENTRENADOR",
                      "MANAGER"
                    ]
                  },
                  "id_creador": {
                    "type": "string",
                    "description": "Staff ID of the HEAD_COACH creating the invitation"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitation created"
          },
          "403": {
            "description": "Only HEAD_COACH can create invitations"
          }
        }
      }
    },
    "/api/invitaciones/equipo/{id}": {
      "get": {
        "summary": "Get all invitations for a team",
        "tags": [
          "Invitations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of invitations"
          }
        }
      }
    },
    "/api/invitaciones/validate/{codigo}": {
      "get": {
        "summary": "Validate invitation code without redeeming",
        "tags": [
          "Invitations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "codigo",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validation result with team info"
          }
        }
      }
    },
    "/api/invitaciones/{id}": {
      "delete": {
        "summary": "Delete/revoke invitation",
        "tags": [
          "Invitations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation deleted"
          }
        }
      }
    },
    "/api/usuarios/{id}": {
      "get": {
        "summary": "Obtener usuario por ID (solo tu propio perfil)",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "403": {
            "description": "No tienes permiso para ver este perfil"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar usuario (solo tu propio perfil)",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado"
          },
          "403": {
            "description": "No tienes permiso para modificar este perfil"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar usuario (solo tu propia cuenta)",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado"
          },
          "403": {
            "description": "No tienes permiso para eliminar este perfil"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/api/usuarios/{id}/password": {
      "patch": {
        "summary": "Actualizar contraseña (solo tu propia cuenta)",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña actualizada"
          },
          "403": {
            "description": "No tienes permiso para cambiar esta contraseña"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    }
  },
...
}